requirements.txt
absl-py==0.11.0
alembic==1.4.3
astunparse==1.6.3
authy==2.2.6
blinker==1.4
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gast==0.3.3
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.18.5
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.26.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
